Move função de tratamento de dados cadastrais para utils
