+++ pages/gestao-usuarios/cadastro/lotes/index.js
@@ -8,6 +8,7 @@ import { CadastrarUsuarioLotes } from '../../../../helpers/RequisicoesConcorrent
 import { colunasValidacaoDadosCadastro } from '../../../../helpers/colunasValidacaoDadosCadastro';
 import { colunasValidacaoRequsicoes } from '../../../../helpers/colunasValidacaoRequisicoes';
 import { redirectHomeGestaoUsuarios } from '../../../../helpers/redirectHome';
+import { Tratamento } from '../../../../utils/tratamentoDadosCadastrais';
 
 export async function getServerSideProps(ctx) {
   const session = await getSession(ctx);
@@ -34,32 +35,6 @@ const colunas = [
 const validarColunas = linha => Object.keys(linha).every(chave => colunas.includes(chave));
 
 const validarColunaLinhas = data => data.every(linha => validarColunas(linha));
-
-const Tratamento = async (data, setDadosReq) => {
-  if (data) {
-    const TratarNome = nome => nome.split(' ').map((item) => item.charAt(0).toUpperCase() + item.slice(1).toLowerCase()).join(' ').trim();
-    const TratarINE = equipe => equipe.toString().replace(/[^0-9]/g, '').padStart(10, '0');
-    const TratarMail = mail => mail.replaceAll(' ', '');
-    const TratarCPF = cpf => cpf.toString().replace(/[^0-9]/g, '');
-    const TratarTelefone = num => num.toString().replace(/[^0-9]/g, '');
-    const dadosTratados = data.map(usuario => {
-      return {
-        'nome': TratarNome(usuario.nome),
-        'equipe': TratarINE(usuario.equipe),
-        'mail': TratarMail(usuario.mail),
-        'cpf': TratarCPF(usuario.cpf),
-        'telefone': TratarTelefone(usuario.telefone),
-        'cargo': usuario.cargo,
-        'municipio_uf': usuario.municipio_uf,
-        'perfil': usuario.perfil,
-        'whatsapp': usuario.whatsapp
-      };
-    });
-    setDadosReq(dadosTratados);
-    return dadosTratados;
-  }
-
-};
 
 const Validacao = async data => {
   const ValidarExistencia = input => input !== null || input !== undefined;
